Red Slime Added
Red Slime prefab added
The animations for the red slime have been created.
A new prefab has been added in place of the enemy with 50 health in the sample scene's fight.
The enemy template section in the script named Spawner inside EnemySpawn has been modified to hold multiple objects, and these templates are assigned based on the value of i in the for loop.